Feature firebase user authentication

+++ server/controllers/chatSession-controller.js
@@ -6,7 +6,7 @@ const httpCreateSession = async (req, res) => {
     res.status(400).json({ "Error Message": "Must provide a session name" });
     return false;
   }
-  const result = await chatSessionModel.createChatSession(req.body.sessionName);
+  const result = await chatSessionModel.createChatSession(req.body.sessionName, req.body.requestedbyUser);
   if (result === false) {
     res.status(500).json({});
     return false;
@@ -15,7 +15,7 @@ const httpCreateSession = async (req, res) => {
 };
 
 const httpGetSessions = async (req, res) => {
-  const result = await chatSessionModel.getChatSessions();
+  const result = await chatSessionModel.getChatSessions(req.body.requestedbyUser);
 
   if (result === false) {
     res.status(500).json({});
@@ -26,6 +26,11 @@ const httpGetSessions = async (req, res) => {
 };
 
 const httpGetSessionHistory = async (req, res) => {
+  if(await chatSessionModel.isAuthorized(req.params.id, req.body.requestedbyUser) === false){
+    res.status(401).json({error: "User is not authorized to access data"});
+    return false;
+  }
+
   const result = await chatSessionModel.getChatSessionChatDetail(req.params.id);
 
   if (result === false) {
@@ -36,6 +41,11 @@ const httpGetSessionHistory = async (req, res) => {
 };
 
 const httpInsertChat = async (req, res) => {
+  if(await chatSessionModel.isAuthorized(req.params.id, req.body.requestedbyUser) === false){
+    res.status(401).json({ error: "User is not authorized to access data"});
+    return false;
+  }
+
   if (!req.body.senderID || typeof req.body.senderID !== "string") {
     res.status(400).json({ error: "Must provide a Sender ID String" });
     return false;
@@ -45,6 +55,8 @@ const httpInsertChat = async (req, res) => {
     res.status(400).json({ error: "Must provide a message String" });
     return false;
   }
+
+
   const result = await chatSessionModel.setChatGlobal(req.params.id, req.body.senderID, req.body.message);
 
   if (result === false) {
@@ -59,11 +71,21 @@ const httpInsertChat = async (req, res) => {
 
 const httpForceBotChat = async (req, res) => {
   // The HTTP response will be send by chatGPT controller.
+  if(await chatSessionModel.isAuthorized(req.params.id, req.body.requestedbyUser) === false){
+    res.status(401).json({ error: "User is not authorized to access data"});
+    return false;
+  }
+
   await chatgptController.generateGPTChat(req.params.id, res);
   return true;
 };
 
 const httpDeleteSession = async (req, res) => {
+  if(await chatSessionModel.isAuthorized(req.params.id, req.body.requestedbyUser) === false){
+    res.status(401).json({ error: "User is not authorized to access data"});
+    return false;
+  }
+
   const result = await chatSessionModel.deleteSession(req.params.id);
   if (result === false) {
     res.status(500).json({});
@@ -73,6 +95,11 @@ const httpDeleteSession = async (req, res) => {
 };
 
 const httpUpdateSession = async (req, res) => {
+  if(await chatSessionModel.isAuthorized(req.params.id, req.body.requestedbyUser) === false){
+    res.status(401).json({ error: "User is not authorized to access data"});
+    return false;
+  }
+
   if (!req.body.sessionName || typeof req.body.sessionName !== "string") {
     res.status(400).json({ error: "Must provide a Session Name String" });
     return false;
@@ -86,6 +113,12 @@ const httpUpdateSession = async (req, res) => {
 };
 
 const httpInsertPersonIntoSession = async (req, res) => {
+  if(await chatSessionModel.isAuthorized(req.params.id, req.body.requestedbyUser) === false){
+    res.status(401).json({ error: "User is not authorized to access data"});
+    return false;
+  }
+
+
   const result = await chatSessionModel.insertPerson(req.params.id, req.params.personid);
   if (result === false) {
     res.status(500).json({});
@@ -95,6 +128,12 @@ const httpInsertPersonIntoSession = async (req, res) => {
 };
 
 const httpRemovePersonFromSession = async (req, res) => {
+  if(await chatSessionModel.isAuthorized(req.params.id, req.body.requestedbyUser) === false){
+    res.status(401).json({ error: "User is not authorized to access data"});
+    return false;
+  }
+
+
   const result = await chatSessionModel.removePerson(req.params.id, req.params.personid);
   if (result === false) {
     res.status(500).json({});
@@ -104,6 +143,12 @@ const httpRemovePersonFromSession = async (req, res) => {
 };
 
 const httpGetPersonsInSession = async (req, res) => {
+  if(await chatSessionModel.isAuthorized(req.params.id, req.body.requestedbyUser) === false){
+    res.status(401).json({ error: "User is not authorized to access data"});
+    return false;
+  }
+
+  
   const result = await chatSessionModel.getPersons(req.params.id);
   if (result === false) {
     res.status(500).json({});
